Azure Functions/requirements.txt
azure-functions
uuid
azure-cosmos
email_validator
python-jose
azure-communication-email
bcrypt
pydantic
pytz